+++ sale-page-et2-maxim-update.php
@@ -39,16 +39,40 @@
                         <div class="block-picture-primary-container">
 
                             <div class="block-container">
-                                <a href="<?php echo site_url('/brands/quoizel-lighting/'); ?>">
+                                <a href="<?php echo site_url('/product-category/quoizel-lighting/quoizel-table-lamps/'); ?>">
                                     <div class="block-picture">
-                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2019/04/LWN5005BN_4_.jpg">', wp_get_attachment_metadata(15942), 15942 ); ?>
+                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2019/05/TF6668VB.jpg">', wp_get_attachment_metadata(22354), 22354 ); ?>
                                     </div>
                                     <div class="block-caption-wrapper">
                                         <div class="block-caption-container">
-                                            <h4 class="block-heading">Quoizel Lighting</h4>
-                                            <h3 class="block-heading">Get 10% off on All of Quoizel Lighting</h3>
-                                            <h4 class="block-heading">Use code "QUOIZELALL10" at checkout!</h4>
-                                            <a style="text-align: center;" href="<?php echo site_url('/brands/quoizel-lighting/'); ?>" class="block-shop-link">Shop Now</a>
+                                            <h4 class="block-heading">Quoizel Table Lamps</h4>
+                                            <h3 class="block-heading">Get 10% off on Quoizel Table Lamps</h3>
+                                            <h4 class="block-heading">Use code "QUOIZEL10" at checkout!</h4>
+                                            <a style="text-align: center;" href="<?php echo site_url('/product-category/quoizel-lighting/quoizel-table-lamps/'); ?>" class="block-shop-link">Shop Now</a>
+                                        </div>
+                                    </div>
+                                </a>
+                            </div>
+
+                        </div>
+
+                    </div>
+
+                    <div class="col-md-4">
+
+                        <div class="block-picture-primary-container">
+
+                            <div class="block-container">
+                                <a href="<?php echo site_url('/product-category/quoizel-lighting/quoizel-floor-lamps/'); ?>">
+                                    <div class="block-picture">
+                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2019/05/TFMK9362VA.jpg">', wp_get_attachment_metadata(22438), 22438 ); ?>
+                                    </div>
+                                    <div class="block-caption-wrapper">
+                                        <div class="block-caption-container">
+                                            <h4 class="block-heading">Quoizel Floor Lamps</h4>
+                                            <h3 class="block-heading">Get 10% off on Quoizel Floor Lamps</h3>
+                                            <h4 class="block-heading">Use code "QUOIZEL10" at checkout!</h4>
+                                            <a style="text-align: center;" href="<?php echo site_url('/product-category/quoizel-lighting/quoizel-floor-lamps/'); ?>" class="block-shop-link">Shop Now</a>
                                         </div>
                                     </div>
                                 </a>